added launching the start page from the user data deleted google sign in

+++ app/src/main/java/ru/wintrade/mvp/view/subscriber/SubscriberMainView.kt
@@ -9,5 +9,5 @@ interface SubscriberMainView : MvpView {
     fun init()
     fun setAvatar(ava: String?)
     fun setName(username: String)
-    fun setSubscriptionCount(count: Long)
+    fun setSubscriptionCount(count: Int)
 }
